[NEST-BE-27] Create API Endpoint for adding likes/undo for userpost, article

+++ src/main/java/com/nest/core/post_management_service/model/Post.java
@@ -54,10 +54,10 @@ public class Post {
     private String type;
 
     @Column(name="likes_count")
-    private int likesCount;
+    private Long likesCount;
 
     @Column(name="view_count")
-    private int viewCount;
+    private Long viewCount;
 
     @Column(name="share_count")
     private int shareCount;
